# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FreshFlow API
  version: 1.0.0
  description: |
    **FreshFlow** is a multi-tenant B2B SaaS platform for fresh produce ordering and warehouse management.
    It connects distributors with restaurant customers through an intelligent ordering system with catch-weight support,
    per-customer pricing, and a complete order lifecycle (DRAFT → SENT → IN_SEPARATION → FINALIZED).

    ## Architecture

    This API uses a hybrid architecture:
    - **tRPC endpoints** (`/trpc/*`) — Type-safe client-server communication via [tRPC v11](https://trpc.io)
    - **REST endpoints** (`/api/v1/*`) — PDF generation, webhooks, Server-Sent Events, and error reporting

    ### tRPC HTTP Convention
    - **Queries** are sent as `GET /trpc/{router}.{procedure}?input={URL-encoded JSON}`
    - **Mutations** are sent as `POST /trpc/{router}.{procedure}` with body `{ "json": <input> }`
    - Responses are wrapped in a tRPC envelope: `{ "result": { "data": <payload> } }`

    ## Authentication

    All authenticated endpoints require a **Supabase JWT** in the `Authorization` header:
    ```
    Authorization: Bearer <supabase_jwt_token>
    ```
    Tokens are obtained by logging in through Supabase Auth on the frontend.

    ## Multi-Tenancy

    FreshFlow uses a tenant/account hierarchy. Most endpoints require context headers:
    - `x-tenant-id` — UUID of the tenant (distributor) for tenant-scoped operations
...

    ## RBAC (Role-Based Access Control)

...

    ## Rate Limiting
...
servers:
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token. Obtain via Supabase Auth login flow.

  # =========================================================================
  # Reusable Parameters
  # =========================================================================
...
  schemas:
...
    OrderStatus:
      type: string
      enum: [DRAFT, SENT, IN_SEPARATION, FINALIZED]
      description: "Order lifecycle: DRAFT → SENT → IN_SEPARATION → FINALIZED"
...
    OrderEvent:
      type: object
      properties:
        type:
          type: string
          enum: [connected, CREATED, UPDATED, FINALIZED, CANCELLED]
        orderId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/OrderStatus"
        timestamp:
          type: string
          format: date-time
        data:
          type: object
          nullable: true

# ===========================================================================
# Paths
# ===========================================================================
paths:
...
  /api/v1/whatsapp/webhook:
    post:
      tags: [Webhooks]
      operationId: whatsappWebhook
      summary: Twilio WhatsApp webhook
      description: |
        Receives incoming WhatsApp messages from Twilio. Processes the message and optionally
        returns a TwiML auto-reply response.

        **Authentication**: None (called by Twilio)
        **Rate Limit**: 1000 requests/minute
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [From, Body]
              properties:
                From:
                  type: string
                  description: Sender phone number (WhatsApp format)
                  example: "whatsapp:+5511999999999"
                Body:
                  type: string
                  description: Message content
                  example: "Quero ver meus pedidos"
      responses:
        "200":
          description: Message processed (TwiML auto-reply or JSON acknowledgement)
          content:
            text/xml:
              schema:
                type: string
                description: TwiML response with auto-reply message
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [received]
        "400":
          description: Missing From or Body fields
        "500":
          description: Webhook processing failed
  /api/v1/orders/events:
    get:
      tags: [Real-Time]
      operationId: orderEventsSSE
      summary: Real-time order updates (Server-Sent Events)
      description: |
        Opens a persistent SSE connection that streams real-time order events to the client.
        Events are filtered based on the user's memberships:
        - Account members receive events for their account's orders
        - Tenant admins receive events for all orders in their tenant

        **Heartbeat**: Every 30 seconds (`:heartbeat {timestamp}`)
        **Authentication**: Bearer JWT required

        ### Event Format
        ```
        data: {"type":"CREATED","orderId":"...","status":"SENT","timestamp":"..."}
        ```
      responses:
        "200":
          description: SSE stream established
          headers:
            Content-Type:
              schema:
                type: string
                enum: [text/event-stream]
            Cache-Control:
              schema:
                type: string
                enum: [no-cache]
            Connection:
              schema:
                type: string
                enum: [keep-alive]
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/OrderEvent"
        "401":
          description: Authentication required
  /api/v1/errors:
    post:
      tags: [Error Reporting]
      operationId: reportFrontendError
      summary: Report a frontend error
      description: |
        Receives error reports from the frontend application. Logs the error to the audit system
        for monitoring and debugging. Authentication is optional — unauthenticated reports are accepted
        but will not be linked to a user.

        **Rate Limit**: 100 requests/minute
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [errorId, message]
              properties:
                errorId:
                  type: string
                  description: Unique error identifier (generated by frontend)
                message:
                  type: string
                  description: Error message
                stack:
                  type: string
                  description: Stack trace (truncated to 2000 chars server-side)
                componentStack:
                  type: string
                  description: React component stack
                url:
                  type: string
                  description: Page URL where error occurred
                userAgent:
                  type: string
                timestamp:
                  type: string
                  format: date-time
      responses:
        "201":
          description: Error report received
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  errorId:
                    type: string
                  message:
                    type: string
                    example: "Error report received"

  # =========================================================================
  # tRPC ENDPOINTS — Auth
  # =========================================================================
...